home page
this is the home page mockup

the “my account” drop down sign in, and drop down nav for quick access
to lists and favorites is mocked up if anyone wants to see.

Examples of list styles (for “most pop recipes” are mocked up if anyone
wants to see that as well.